requirements.txt
python-dotenv
pandas
numpy
scikit-learn
dill
PyYAML
mlflow
fastapi
uvicorn
python-multipart
